# Conflicts:
#	nifi-docs/src/main/asciidoc/administration-guide.adoc
#	nifi-nar-bundles/nifi-framework-bundle/nifi-framework/nifi-resources/src/main/resources/bin/nifi-env.sh
#	nifi-nar-bundles/nifi-framework-bundle/nifi-framework/nifi-web/nifi-web-security/src/main/java/org/apache/nifi/web/security/oidc/OidcService.java
#	nifi-nar-bundles/nifi-framework-bundle/nifi-framework/nifi-web/nifi-web-security/src/main/java/org/apache/nifi/web/security/oidc/StandardOidcIdentityProvider.java
#	nifi-nar-bundles/nifi-framework-bundle/nifi-framework/nifi-web/nifi-web-security/src/test/groovy/org/apache/nifi/web/security/oidc/StandardOidcIdentityProviderGroovyTest.groovy

+++ nifi-nar-bundles/nifi-standard-services/nifi-record-serialization-services-bundle/nifi-record-serialization-services/src/test/java/org/apache/nifi/json/TestInferJsonSchemaAccessStrategy.java
@@ -182,7 +182,7 @@ public class TestInferJsonSchemaAccessStrategy {
         assertEquals(2, possibleTypes.size());
         assertTrue(possibleTypes.contains(RecordFieldType.STRING.getDataType()));
         assertTrue(possibleTypes.contains(RecordFieldType.ARRAY.getArrayDataType(RecordFieldType.CHOICE.getChoiceDataType(
-                RecordFieldType.LONG.getDataType(),
+                RecordFieldType.INT.getDataType(),
                 RecordFieldType.STRING.getDataType()
         ))));
